[REFACT] 지역, 시군구 리팩토
