+++ src/main/java/konkuk/tourkk/chons/domain/areasigungu/infrastructure/AreaRepository.java
@@ -1,7 +1,7 @@
-package konkuk.tourkk.chons.domain.festival.infrastructure;
+package konkuk.tourkk.chons.domain.areasigungu.infrastructure;
 
 import java.util.Optional;
-import konkuk.tourkk.chons.domain.festival.domain.entity.Area;
+import konkuk.tourkk.chons.domain.areasigungu.domain.entity.Area;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AreaRepository extends JpaRepository<Area, Long> {
