--- src/main/java/konkuk/tourkk/chons/domain/festival/application/res/AreaResponse.java
@@ -1,11 +1,0 @@
-package konkuk.tourkk.chons.domain.festival.application.res;
-
-import lombok.Getter;
-
-@Getter
-public class AreaResponse {
-
-    String code;
-
-    String name;
-}
