+++ src/main/java/konkuk/tourkk/chons/domain/house/presentation/controller/HouseController.java
@@ -4,11 +4,11 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import konkuk.tourkk.chons.domain.house.application.HouseService;
-import konkuk.tourkk.chons.domain.house.application.apiresponse.AreaListResponse;
+import konkuk.tourkk.chons.domain.areasigungu.application.dto.res.AreaListResponse;
 import konkuk.tourkk.chons.domain.house.presentation.dto.req.HouseListRequest;
 import konkuk.tourkk.chons.domain.house.presentation.dto.req.HouseRequest;
 import konkuk.tourkk.chons.domain.house.presentation.dto.res.HouseResponse;
-import konkuk.tourkk.chons.domain.sigungu.application.AreaSigunguService;
+import konkuk.tourkk.chons.domain.areasigungu.application.service.AreaSigunguService;
 import konkuk.tourkk.chons.domain.user.domain.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
