--- src/main/java/konkuk/tourkk/chons/domain/sigungu/application/res/SigunguResponse.java
@@ -1,13 +1,0 @@
-package konkuk.tourkk.chons.domain.sigungu.application.res;
-
-import lombok.Getter;
-
-@Getter
-public class SigunguResponse {
-
-    Long rnum;
-
-    String code;
-
-    String name;
-}
